+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 
+- Changed logger to catch an exception from a writer, not to interrupt and call the next writer if available.
+
 ### Deprecated
 
 ### Removed